+++ src/main/java/heavysnow/heath/controller/LoginController.java
@@ -1,8 +1,7 @@
 package heavysnow.heath.controller;
 
 
-import heavysnow.heath.common.LoginMemberHolder;
-import heavysnow.heath.dto.LoginResponseDto;
+import heavysnow.heath.dto.login.LoginResponseDto;
 import heavysnow.heath.service.LoginService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
@@ -10,8 +9,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Optional;
 
 @RestController
 @RequiredArgsConstructor
