+++ src/main/java/heavysnow/heath/dto/login/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package heavysnow.heath.dto;
+package heavysnow.heath.dto.login;
 
 
 import lombok.AccessLevel;
